requirements.txt
Streamlit 
python-dotenv 
openai 